Remove unzip for Pully types
The reason is that although it type makes it seem that it "extracts" two vectors from one, in fact
it just duplicates the vector and applies fst/snd to the elements. Since this behavior is a bit
surprising, it's better that the user has to do this manually when needed.